refactor: NewCardRouteScreen '카드 추가 - 화면 이동' 테스트 개선
